Refactor code into MVC model
